rewrote TwoDerivedExplicitDecoration.cpp to c++
