+++ CMakeLists.txt
@@ -15,5 +15,5 @@ add_executable(AwesomeBugs
         code/JavaRecursionExamples/mutualrec/non_term/factory/c++/TODO_in_c++_Factory.cpp
         code/JavaRecursionExamples/mutualrec/non_term/long_chain/c++/TODO_in_c++_LongChain.cpp
         code/JavaRecursionExamples/mutualrec/non_term/two_derived/c++/TODO_in_c++_TwoDerived.cpp
-        code/JavaRecursionExamples/mutualrec/non_term/two_derived_explicit_decoration/c++/TODO_in_c++_TwoDerivedExplicitDecoration.cpp
-        code/test/c++/TODO_in_c++_hello_world.cpp)
+        code/JavaRecursionExamples/mutualrec/non_term/two_derived_explicit_decoration/c++/TwoDerivedExplicitDecoration.cpp
+        code/test/c++/hello_world.cpp)
