--- code/test/c++/TODO_in_c++_hello_world.cpp
@@ -1,6 +1,0 @@
-class HelloWorld {
-    public static void main(String[] args) {
-        System.out.println("Hello World!");
-        int a = 5 / 0;
-    }
-}
