+++ code/test/c++/hello_world.cpp
@@ -1,0 +1,6 @@
+#include "iostream"
+    int main() {
+        std::cout << "Hello World!";
+        int a = 5 / 0;
+        return 0;
+    }
